Health bar and stamina bar added
Also took care of an issue with the movement system where if you ran out of stamina you would still move quick, but without the running animation.
Added a player info script that I think is going to help a lot.
The player info script is a way for these components to grab information about the player such as what state it is in.